feat: create new function deepClone
deepClone is a function that can be called as a module in order to generate a new array, object or ordinary variable's copy. 
When we are dealing with array and object types, differently than other data types, we refer directly to a memory position. 
In order to prevent unwanted changes, deepClone generates a new variable, even if it's an array or object, with the same content, but in
a different memory position for each content value.